requirements.txt
python-can
opencv-python
pytesseract
numpy
scapy